adjusted the authorization policy around discussions to ensure the communication authorization is properly set + cascaded to discussions
